requirements.txt
mlflow #==2.9.2
apache-airflow #==2.8.1  # Ou utilisez Docker pour simplicité
scikit-learn #==1.3.2
pandas #==2.1.4
numpy #==1.25.2
semantic-kernel #==0.9.0b1  # Pour agents LLM (besoin d'OpenAI API key)
openai #==1.3.7  # Pour LLM (ex. : GPT-4 via Semantic Kernel)
pyyaml #==6.0.1  # Pour config.yaml
pytest #==7.4.3  # Pour tests
ollama #==0.1.7

# Testing & CI/CD (DeepSeek's)
pytest-asyncio #==0.21.0
pytest-cov #==4.1.0
flake8 #==6.0.0
black #==23.7.0
codecov #==2.1.12
bandit #==1.7.5
